+++ Model/DeletedData.php
@@ -7,7 +7,7 @@ use Carbon\CarbonInterface;
 class DeletedData
 {
     public function __construct(
-        public int $id,
-        public CarbonInterface $deletedDate,
+        public ?int $id,
+        public ?CarbonInterface $deletedDate,
     ) {}
 }