+++ Model/TimesheetData.php
@@ -8,13 +8,14 @@ readonly class TimesheetData
 {
     public function __construct(
         public int $id,
-        public int $ticketId,
-        public int $projectId,
+        public ?int $ticketId,
+        public ?int $projectId,
         public ?string $description,
         public float $hours,
-        public string $username,
+        public ?int $userId,
+        public ?string $username,
+        public ?string $kind,
         public ?CarbonInterface $workDate = null,
         public ?CarbonInterface $modified = null,
-        public string $kind,
     ) {}
 }